requirements.txt
asgiref==3.8.1
autopep8==2.2.0
bakery==1.0.0
Django==5.0.6
django-extensions==3.2.3
django-ordered-model==3.7.4
exceptiongroup==1.2.2
flake8==7.0.0
iniconfig==2.0.0
mccabe==0.7.0
model-bakery==1.18.2
mysqlclient==2.2.4
packaging==24.1
pluggy==1.5.0
pycodestyle==2.11.1
pyflakes==3.2.0
pytest==8.2.2
python-dotenv==1.0.1
sqlparse==0.5.0
tomli==2.0.1
typing_extensions==4.12.1
